% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLong.R
\docType{data}
\name{dataLong}
\alias{dataLong}
\title{Simulated multiple longitudinal data}
\format{
A data frame which contains id, obstime, x1, x2, and Y1,...,Y10.
\describe{
  \item{id}{patients identifier}
  \item{obstime}{observed times for longitudinal measerements}
  \item{Y1}{the first longitudinal response}
  \item{Y2}{the second longitudinal response}
  \item{Y3}{the third longitudinal response}
  \item{Y4}{the fourth longitudinal response}
  \item{x1}{a binary covariate}
  \item{x2}{a continuous covariate}
}
}
\usage{
dataLong
}
\description{
Simulated data were generated in which "Y1",...,"Y10" follows a Gaussian distribution, "obstime" represents the observed time with some covariates.
}
\seealso{
\code{\link{UJM}}
}
\keyword{datasets}
