% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\name{DP}
\alias{DP}
\title{Dynamic prediction type 1}
\usage{
DP(
  object,
  s = s,
  t = t,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  quiet = FALSE,
  dataLong,
  dataSurv
)
}
\arguments{
\item{object}{an object inheriting from class TS}

\item{s}{the landmark time for prediction}

\item{t}{the window of prediction for prediction}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Dynamic prediction for TSJM
}
\details{
Estimate DP for joint modeling based on TS approach.
}
\examples{
\dontrun{
rm(list=ls())
data(dataLong)
data(dataSurv)
set.seed(2)
INDTRAIN <- sample(dataSurv$id, 0.8 * (dim(dataSurv)[1]))
INDVALID <- dataSurv$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong,
  dataLong$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv,
  dataSurv$id \%in\% INDTRAIN
)

dataLong_v <- subset(
  dataLong,
  dataLong$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv,
  dataSurv$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, death) ~ x1 + x2
model <- list("intercept", "linear", "quadratic")

TS0 <- TS(formFixed, formRandom, formGroup, formSurv,
         nmark = 3, K1 = 15, K2 = 15,
         model = model, n.chains1 = 1, n.iter1 = 20, n.burnin1 = 10,
         n.thin1 = 1,  n.chains2 = 1, n.iter2 = 30, n.burnin2 = 10,
         n.thin2 = 1, simplify=TRUE, Obstime = "obstime", ncl = 2,
         DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

model <- list("linear", "linear", "linear")

TS1 <- TS(formFixed, formRandom, formGroup, formSurv,
         nmark = 3, K1 = 15, K2 = 15,
         model = model, n.chains1 = 1, n.iter1 = 20, n.burnin1 = 10,
         n.thin1 = 1,  n.chains2 = 1, n.iter2 = 30, n.burnin2 = 10,
         n.thin2 = 1, simplify=TRUE, Obstime = "obstime", ncl = 2,
         DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)



DP <- DP(TS1,
         s = 0.5, t = 0.5, n.chains = 1, n.iter = 20, n.burnin = 1000,
         n.thin = 1,
         DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)


}




}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
