% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSJM.R
\docType{package}
\name{TSJM}
\alias{TSJM-package}
\alias{TSJM}
\title{TSJM Package}
\description{
Employing a two-stage strategy for variable selection in joint modeling of multiple longitudinal measures and survival outcomes. The initial stage involves estimating K one-marker joint models for survival outcome and each marker. In the second stage, time-varying covariates are incorporated into a proportional hazard model to estimate parameters. Both phases follow the Bayesian paradigm. Additionally, this approach facilitates the computation of dynamic predictions based on the estimated parameter values.
}
\references{
\enumerate{
\item
Baghfalaki, T., Hashemi, R., Catherine Helmer & Jacqmin-Gadda, H. (2024). A Two-stage Joint Modeling Approach for Multiple Longitudinal Markers and Time-to-event Data. Submitted.

}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
\keyword{TSJM}
